+++ src/store/peer/peerActions.ts
@@ -3,6 +3,9 @@ import { Dispatch } from "redux";
 import { DataType, PeerConnection } from "../../helpers/peer";
 import { addConnectionList, removeConnectionList } from "../connection/connectionActions";
 import { addConnectionRequest } from "../connection/connectionRequestActions";
+import { createLogger } from "../../services/logService";
+
+const log = createLogger('PeerActions')
 
 export const startPeerSession = (id: string) => ({
     type: PeerActionType.PEER_SESSION_START, id
@@ -18,19 +21,21 @@ export const setLoading = (loading: boolean) => ({
 
 export const startPeer: () => (dispatch: Dispatch) => Promise<void>
     = () => (async (dispatch) => {
+        log.info('Starting peer session')
         dispatch(setLoading(true))
         try {
             const id = await PeerConnection.startPeerSession()
             PeerConnection.onIncomingConnection((conn) => {
                 const peerId = conn.peer
-                console.log("Incoming connection: " + peerId)
+                log.info('Incoming connection from peer: ' + peerId)
                 // Don't auto-accept - dispatch a connection request instead
                 dispatch(addConnectionRequest(peerId))
             })
+            log.debug('Peer session started successfully with ID: ' + id)
             dispatch(startPeerSession(id))
             dispatch(setLoading(false))
         } catch (err) {
-            console.log(err)
+            log.error('Failed to start peer session', err)
             dispatch(setLoading(false))
         }
     })